# restaurant-pos-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Tamanna Restaurant POS API
  description: >
    Comprehensive REST API documentation for Tamanna Restaurant Pure Vegetarian POS & ERP System.
    Includes Point of Sale, KDS, Dynamic Tables, Razorpay Online Gateway, SMS/OTP, Automated Cron Scheduling,
    Excel/PDF document exports, and Real-time WebSockets.
  version: 1.0.0
  contact:
    name: Tamanna Restaurant IT Support
    email: support@tamannarestaurant.com
servers:
  - url: http://localhost:5000
    description: Local Development & POS Server
paths:
...
  /api/reports/export/sales-excel:
    get:
      summary: Export sales and tax breakdown to Microsoft Excel (.xlsx)
      tags:
        - Reports & Exports
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Multi-sheet Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
  /api/reports/export/inventory-excel:
    get:
      summary: Export current inventory stock & valuation to Microsoft Excel (.xlsx)
      tags:
        - Reports & Exports
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Multi-sheet Excel workbook
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT